Gave component a more generic name of Date. Removed moment.js, now using pre-existingtimeago.js package
