+++ Source/Plugins/Core/com.equella.core/js/package.json
@@ -44,7 +44,6 @@
     "lodash": "4.17.15",
     "luxon": "1.16.0",
     "material-ui-pickers": "2.2.4",
-    "moment": "^2.26.0",
     "oeq-cloudproviders": "git+https://github.com/apereo/openEQUELLA-cloudprovidersdk.git#32d958ddfff64ca748e7e1b2eae0f0487946a487",
     "prop-types": "15.7.2",
     "react": "16.13.1",
